refactor: move notifications storage logic into useNotificationsStore for better encapsulation
