+++ store/useNotificationsStore.ts
@@ -1,6 +1,23 @@
 import { create } from 'zustand';
 
-import { notificationsStorage } from './notifications-storage';
+import { miscStorage } from './misc-storage';
+
+const NOTIFICATIONS_LAST_VISITED_KEY = 'notifications-last-visited';
+
+const notificationsStorage = {
+  setLastVisited: (timestamp: number) => {
+    miscStorage.set(NOTIFICATIONS_LAST_VISITED_KEY, timestamp);
+  },
+
+  getLastVisited: (): number | null => {
+    const timestamp = miscStorage.getNumber(NOTIFICATIONS_LAST_VISITED_KEY);
+    return timestamp ?? null;
+  },
+
+  clearLastVisited: () => {
+    miscStorage.delete(NOTIFICATIONS_LAST_VISITED_KEY);
+  },
+};
 
 interface NotificationsState {
   lastVisited: number | null;
